Move backend to platform and wait for connection in timer service
